requirements.txt
Flask-SQLAlchemy==3.0.5
Flask==2.2.5
psycopg2-binary==2.9.9